+++ cmd/tofu/provider_source.go
@@ -15,25 +15,12 @@ import (
 
 	"github.com/apparentlymart/go-userdirs/userdirs"
 	"github.com/hashicorp/terraform-svchost/disco"
-	orasRemote "oras.land/oras-go/v2/registry/remote"
-	orasAuth "oras.land/oras-go/v2/registry/remote/auth"
 
 	"github.com/opentofu/opentofu/internal/addrs"
 	"github.com/opentofu/opentofu/internal/command/cliconfig"
-	"github.com/opentofu/opentofu/internal/command/cliconfig/ociauthconfig"
 	"github.com/opentofu/opentofu/internal/getproviders"
-	"github.com/opentofu/opentofu/internal/httpclient"
 	"github.com/opentofu/opentofu/internal/tfdiags"
 )
-
-// ociCredsPolicyBuilder is the type of a callback function that the [providerSource]
-// functions will use if any of the configured provider installation methods
-// need to interact with OCI Distribution registries.
-//
-// We represent this indirectly as a callback function so that we can skip doing
-// this work in the common case where we won't need to interact with OCI registries
-// at all.
-type ociCredsPolicyBuilder func(context.Context) (ociauthconfig.CredentialsConfigs, error)
 
 // providerSource constructs a provider source based on a combination of the
 // CLI configuration and some default search locations. This will be the
@@ -276,78 +263,4 @@ func providerDevOverrides(configs []*cliconfig.ProviderInstallation) map[addrs.P
 	// the validation logic in the cliconfig package. Therefore we'll just
 	// ignore any additional configurations in here.
 	return configs[0].DevOverrides
-}
-
-// getOCIRepositoryStore instantiates a [getproviders.OCIRepositoryStore] implementation to use
-// when accessing the given repository on the given registry, using the given OCI credentials
-// policy to decide which credentials to use.
-func getOCIRepositoryStore(ctx context.Context, registryDomain, repositoryName string, credsPolicy ociauthconfig.CredentialsConfigs) (getproviders.OCIRepositoryStore, error) {
-	// We currently use the ORAS-Go library to satisfy the [getproviders.OCIRepositoryStore]
-	// interface, which is easy because that interface was designed to match a subset of
-	// the ORAS-Go API since we had no particular need to diverge from it. However, we consider
-	// ORAS-Go to be an implementation detail here and so we should avoid any ORAS-Go
-	// types becoming part of the direct public API between packages.
-
-	// ORAS-Go has a bit of an impedence mismatch with us in that it thinks of credentials
-	// as being a per-registry thing rather than a per-repository thing, so we deal with
-	// the credSource resolution ourselves here and then just return whatever we found to
-	// ORAS when it asks through its callback. In practice we only interact with one
-	// repository per client so this is just a little inconvenient and not a practical problem.
-	credSource, err := credsPolicy.CredentialsSourceForRepository(ctx, registryDomain, repositoryName)
-	if ociauthconfig.IsCredentialsNotFoundError(err) {
-		credSource = nil // we'll just try without any credentials, then
-	} else if err != nil {
-		return nil, fmt.Errorf("finding credentials for %q: %w", registryDomain, err)
-	}
-	client := &orasAuth.Client{
-		Client: httpclient.New(), // the underlying HTTP client to use, preconfigured with OpenTofu's User-Agent string
-		Credential: func(ctx context.Context, hostport string) (orasAuth.Credential, error) {
-			if hostport != registryDomain {
-				// We should not send the credentials we selected to any registry
-				// other than the one they were configured for.
-				return orasAuth.EmptyCredential, nil
-			}
-			if credSource == nil {
-				return orasAuth.EmptyCredential, nil
-			}
-			creds, err := credSource.Credentials(ctx, ociCredentialsLookupEnv{})
-			if ociauthconfig.IsCredentialsNotFoundError(err) {
-				return orasAuth.EmptyCredential, nil
-			}
-			if err != nil {
-				return orasAuth.Credential{}, err
-			}
-			return creds.ToORASCredential(), nil
-		},
-		Cache: orasAuth.NewCache(),
-	}
-	reg, err := orasRemote.NewRegistry(registryDomain)
-	if err != nil {
-		return nil, err // This is only for registryDomain validation errors, and we should've caught those much earlier than here
-	}
-	reg.Client = client
-	err = reg.Ping(ctx) // tests whether the given domain refers to a valid OCI repository and will accept the credentials
-	if err != nil {
-		return nil, fmt.Errorf("failed to contact OCI registry at %q: %w", registryDomain, err)
-	}
-	repo, err := reg.Repository(ctx, repositoryName)
-	if err != nil {
-		return nil, err // This is only for repositoryName validation errors, and we should've caught those much earlier than here
-	}
-	// NOTE: At this point we don't yet know if the named repository actually exists
-	// in the registry. The caller will find that out when they try to interact
-	// with the methods of [getproviders.OCIRepositoryStore].
-	return repo, nil
-}
-
-// ociCredentialsLookupEnv is our implementation of ociauthconfig.CredentialsLookupEnvironment
-// used when resolving the selected credentials for a particular OCI repository.
-type ociCredentialsLookupEnv struct{}
-
-var _ ociauthconfig.CredentialsLookupEnvironment = ociCredentialsLookupEnv{}
-
-// QueryDockerCredentialHelper implements ociauthconfig.CredentialsLookupEnvironment.
-func (o ociCredentialsLookupEnv) QueryDockerCredentialHelper(ctx context.Context, helperName string, serverURL string) (ociauthconfig.DockerCredentialHelperGetResult, error) {
-	// TODO: Implement this
-	return ociauthconfig.DockerCredentialHelperGetResult{}, fmt.Errorf("support for Docker-style credential helpers is not yet available")
 }
